backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
supabase
httpx
networkx
qrcode[pil]
Pillow
pandas
python-dotenv
pydantic
websockets
python-socketio
aiofiles
requests
numpy
